solved Bill Gates' "no one wants more than 640k of ram" problem and changed iteration_counter from int to unsigned long. No really I tough I was the only one to use it and I didn't mind that it was overflowing from time to time... anyway now it's solved, enjoy!

+++ emilibase.cpp
@@ -200,7 +200,7 @@ static inline void setTimer(int maxTime)
 /*
  * Iteration counter
  */
-static int iteration_counter_ ;
+static unsigned long iteration_counter_ ;
 
 void emili::iteration_counter_zero()
 {
